icb: the Metal half of the indirect command buffer moves out from under the decode
`runtime/icb/mod.rs` was two modules in one file: the guest's own
serialization — the type-7 create body, the command layout, the encode and
decode of one command slot's bytes, all of it rail-neutral by construction —
and seventeen items that name `MTL*` objects, each carrying its own
`#[cfg(all(feature = "backend-metal", target_os = "macos"))]` and
interleaved through the first half.

Interleaved, the boundary was unreadable. Nothing in the text of two
neighbouring functions said which of them the Vulkan arm compiles; the
answer was an attribute several screens up, and it had to be re-derived on
every read. `runtime::draw` already made this split, and `compute_exec`
has `metal.rs`/`vulkan.rs` beside its neutral parent — this file was the
outlier.

So the Metal half becomes `runtime/icb/metal.rs`: materialization, the
per-`(task, ref)` host ICB cache, the render and compute slot fills, the
vertex descriptor a filled render slot needs, and the compute PSO that has
to be built with `supportIndirectCommandBuffers`. The module carries the
gate once and no item inside it repeats one; `icb/mod.rs` keeps exactly one
`cfg`, on the `mod` declaration itself, which is the lawful question.

Pure placement — no item's body changed. Two incidental cleanups came with
it: `clear_icb_cache` now asks the Metal module to clear its own map rather
than reaching for a static it does not own, and `materialize_metal_icb`
loses a `let _ = TYPE7_OBJECT_ICB;` that held an import alive for nothing.

metal 1397, vulkan 2181, both 2265 — unchanged; matrix, fmt, clippy clean,
and no new rustdoc warnings on either arm.
